Preserve old semantics for run-action concerning wait
The wait flag allows you to specify no time arguments for run action,
which then allows you to wait forever (it prevents a timer running).
Also passing `wait=true` is semantically the same.

The problem with the last changes is it prevented the use of this and
error'd out with the timeout being too small. The fix is to expose if
the flag is in the forever mode.