Raft lease extension timeout
The following changes allows extensions on leases to have a larger
timeout when attempting to enqueue on to the raft lease queue. The aim
behind this change is to allow less churn for a given application lease.

During load testing it is seen that when claims block the queue all
subsequent extensions are also thrown away. This is highly problematic
as it causes application leader changes, which also cause more churn in
the database. The change here is to be more forgiving for extensions and
hope after a lost claim, we get to push some extensions through so we
can handle more throughput.