Raft lease retries with exceeded dropped
The following ensures that we identify the error when we can't enqueue
onto the raft lease queue. If that occurs we shouldn't retry for claims,
but we do want to retry for extensions, but only a limited number of
times.

This is because when performing load testing we identified a problem
where we retried enqueuing command failures. This caused issues when
attempting to claim or extend in subsequent calls. Why attempt to claim
and extend if we know the deadline will already be exceeded. We should
fail fast and fail early.

The changes here just expose that error instead of ErrDropped and use
ErrDeadlineExceeded. That way we can identify that error and better deal
with those situations.