Serialise raftlease command until a later stage
To enable better optimisations in op queue, we need to know what type
the command operation is, along with the lease holder. Constantly
leaking this out of the command into the api params requests was
beginning to look like a hack. Instead just send the whole command up
the chain into the queue so it can inspect the whole command.

The code is actually simple, yet it caused lots of churn in the api and
the apiserver. A full round of testing is required to ensure that all
cases are working. It's a shame that we had to bounce the facade, but
that is life.

More optimisations in the future should be possible. Using the lease
holder to push priority queues into operations etc.