Apply synchronously, response asynchronous
The following changes how apply works in a subtle way. Instead of
applying and waiting for apply log and response synchronously, we split
it into two phases. The first phase is synchronous, applying the log and
putting back pressure onto the caller and the queue. Waiting for the
raft response is then done asynchronously, so we don't have to wait for
the flush to disk.

The second phase is done in a similar way to how the pubsub queue is
implemented. The idea behind this change is to give enough back pressure
to the client, but not so much that we're now waiting for raft and mongo
to finish writing to the disk.

The potential downside to this, is a spike in goroutines.