Ensure we only expire leases with the same holder
The following ensures that we don't attempt to expire leases that we
don't hold the lease for. Ensuring that we don't expire leases that we
don't hold should prevent any races that could happen if multiple
expiries for the same lease key are requested but have a different
holder.

The code attempts best effort, but the code does log warnings if there
are some really unpredictable behaviour.

More tests are required to ensure that we do actually do what we expect,
but the existing tests correctly pass. New tests will follow in another
PR.