Fixes re-downloading of caas charm on restart.
When a podspec operator restarts it fails to re download the charm when it sits in volatile storage. This is because the caasoperator controls the application update channel to the uniter. This change removes this control from the caasoperator and resolves the same way as IAAS and sidecar charms by watching application state from the controller.