Update potential failures in destroy_model.
If the controller is unreachable, the word error is not used in output
Instead check the output for "Unable to get the model status from the API"
from "Unable to get the model status from the API: no reachable servers"
and "Unable to get the model status from the API: connection is shut
down". Watch for erroneous failures in the future.