Dependency engine worker start metrics
The following exposes the number of starts a worker has for a given
model. To reduce cardinality the metrics are split into the following:

 1. Number of starts per worker globally.
 2. Number of starts of workers by model.

We do it this way to prevent the explosion of larger controllers with
lots of models dealing with lots of metrics. Hopefully these metrics
will then help operators of Juju see if there are problems with
operating Juju from things like grafana.