Update provisioner_task test call count expectations
Due to the recent changes in the way that the provisioner_task processes
watcher notifications, it is likely that in resource-constrained systems
(CI or when running with -race) that the provisioner task main loop will
be triggered more than once in some tests.

As a result, the code that fetches the latest AZ distribution group
information may be invoked more than once. This commit makes the AZ
tests more robust by updating the relevant test expectations to account
for such scenarios.