Thread the DB through the API server
It should be now possible to pick up the database within the facade
calls. From here we should look into wiring up the new overlord package
to manage state.

We should swap out the *sql.Tx for a interface at this level, so we can
use both unit and integration tests.