Reduce API server metric cardinality
The number of API server metric quantile permutations are growing
because the rpc.ErrorCode can lead to too many deviations. The fix for
now is to reduce this and only compute success or errors. This way we
can be sure that the cardinality can be finite.

This should reduce the total number of memory used for every API server
instance.