Do not force "stable" channel when re-deploying existing local charms
When deploying local charms, they do not get associated with a
particular channel as this information is not available at deploy time.

When deploying a bundle containing local charms multiple times, we
should never force the channel for local charms as this prevents Juju
from locating the already-existing local charm causing the error
described in LP1954933.