Remove txn watcher wrench
The following removes the txn watcher wrench. It calls too frequently in
production applications. This removes it completely, but I'm not adverse
to either creating a controller/model config for enabling wrenches
(calling wrench.SetEnable method), similar to the setup for logging
config. Alternatively, just throttling the existence of the wench every
minute or so.

Either way, we just need to back off calling this.