add user-agent to Equinix provider
Adds a user-agent to API requests to Equinix.  This user-agent follows
the prefix format in-use by the Azure provider.  The suffix of the
user-agent is the default packngo user-agent (expressing the version of
packngo in use).

From an Equinix Metal perspective, this will help gauge Juju consumption
of the Equinix Metal API.

Signed-off-by: Marques Johansson <mjohansson@equinix.com>