Listen on a random port when detecting the local dqlite address
Prior to this change, the local address detection code would listen on
the same port as dqlite. Due to the way that the linux kernel handles
closed socket (remain in WAIT state until the default SO_LINGER timeout
expires), the original approach would prevent the dqlite code from
binding to the port as it is not setting the SO_REUSEPORT/SO_REUSEADDR
flags on the socket.

Instead of manually tuning SO_LINGER ourselves (which would make the
code non-portable), this commit changes the logic to simply use a random
available port to work around the kernel's limitation.