Refactor dbaccessor worker
The worker is now refactored to allow for testing and building with cgo
and no-cgo implementations. This should drastically improve workflow. It
also ensures that the cgo changes do not leak outside of this folder for
now.

More work is require for the repl, which will following in other
commits.