[JUJU-316] DQLite auto-cluster in ha mode with TLS support (#13617)
* Support TLS and auto-clustering for dqlite instances

This commit refactors and enhances the previous implementation so that
we can automatically establish the dqlite cluster based on the
information that we can obtain from the agent's configuration.

In particular, the code will now leverage the (generated) CA and
controller certificates used by the controller to encrypt the traffic
between the dqlite instances. In addition, we rely on client certificate
authentication to prevent unauthorized clients from connecting to
dqlite.

Another aspect of the refactoring work has to do with the selection of a
suitable address for dqlite to listen to and advertise to other peers.
This is achieved by scanning the machine's NIC addresses and selecting
the first available cloud-local address. By avoiding public addresses we
ensure that DB traffic is always routed internally (important on public
clouds where public traffic may incur additional costs).

The last bit of the puzzle is the detection of usable remote peers. We
setup a TCP listener on a random port that serves a random nonce. Then
we iterate the list of API addresses, filter out any address that does
not hava a cloud-local scope (see above), extract the host part of the
IP and try to connect to the nonce port on that IP (read timeouts ensure
we bail out fast if the nonce server is not reachable) and read back the
nonce.

This allows auto-clustering to work because the agent configuration for
the HA nodes is guaranteed to contain the cloud-local API address for
the original controller. As a result, the newly added HA machines can
always reach the dqlite server on the original controller and discover
the other peers via RAFT.

* Quote lxc exec commands with pipes in dqlite Makefile