[JUJU-430] Refactor DBAccessor Worker (#13613)
* Refactor dbaccessor worker

The worker is now refactored to allow for testing and building with cgo
and no-cgo implementations. This should drastically improve workflow. It
also ensures that the cgo changes do not leak outside of this folder for
now.

More work is require for the repl, which will following in other
commits.

* Ensure passing of NewApp is via the manifold

To allow for better dependency management to enable better testing, we
pass the DBApp construction through the manifolds.

* Return REPL from a DBApp

Depending on the DBApp it will return a REPL based on that. For noop
DBApp returns not supported, it probably should just return nil for the
error and just ignore accessing the REPL.

* Rename noop app to unsupported app

As it's no longer just performing non-operations, it's actually erroring
out when calling a method, we should name the struct accordingly.