Sync cache model unit set charm url
The following forces a cache sync change during a SetCharmURL from the
uniter API request. The current implementation waits for the unit in the
model cache to reflect the change. If the change never happens, then a
timeout will be recorded and the process will start again.

The following code says, that if the change was dropped during waiting
for the change, then sync the model change directly. This should then
force the URL to be the correct value and other items can then query the
data in a correct manor.