Engine Report: Model Cache
The following improves the reporting for the model cache. Currently,
it's not possible to know what the current state of the model cache is
for a given controller.

In an ideal world we would have an API to dump the model cache, except
that falls down for HA environments. The correct way would be to query
every controller in the fleet to report back their current information.
That's a bit of work and something that can be done in the future, but
for now, we can piggy-back off the juju_engine_report to dump out more
information.

We don't dump all the information, just enough to help diagnose
potential problems with in the eventual consistent cache.