Retry ping before failing
The following attempts to retry a ping before failing. This should
improve the transient errors we can get if a ping on a session fails.
Failure of a ping causes a restart of the API server. The interesting
part is that we know the session to mongo is ok, as mongo cursor
iterators will fail immediately in .Next() calls and the .Close() will
return the error. The latter should cause the apiserver to restart.

The latter case can be replicated easily by putting juju into HA mode
and then calling rs.stepDown(10) from the mongo shell. At no point will
the .Ping() fail, instead other session requests will fail.

Additionally, a modelStateWorker is only required on a controller NOT
for every model. The mongo is the same mongo for the whole controller,
so the worker is doing busy work for no reason.