Remove fake NIC generation fallback logic from instancepoller worker
Now that all Juju providers (except manual) implement the
NetworkInterfaces method we can finally remove the fallback logic within
the instancepoller worker that would otherwise generate a list of fake
NICs to emulate the output of NetworkInterfaces.

The prior logic has been replaced with a check that outputs a warning to
the logs if the instancepoller ever encounters a provider that does not
implement this method. This is meant as a hint to folks working on
adding support for new providers in the future that they must implement
the method.

Note that the instancepoller ignores manual machines so this particular
call does not need to be implemented for that provider.