Re-write transaction builder
The transaction builder now runs all transaction runners on commit,
which then allows retries much more simpler as we can get transient
network issues.

-----------------------------------------------------------------------

Export CGO_ENABLED=0

The following exports CGO_ENABLED=0 in the make file so we don't have to
change every build target. At a later date we can update this to be
specific, but for now brute force it.