Remove acceptancetests
We're spending too long on old testing infrastructure and not spending
time on working on real bugs we have in Juju. People keep adding and
changing this code, but it's deprecated and dead.

All this code is available in the history of the project. As these tests
are dead weight, we should look to the future and write new tests.
Either they be integration tests in bash or in go.