Charmhub refresh results are unordered
The following updates the charmhub refresh API clients so that all
results are correctly ordered. This was written in the documentation of
the charmhub client API, but it was not observed when using the client.
This can be seen using the charmrevisionupdater expects all results to
be in order. If the results are out of order we end up creating place
holders with invalid revisions.

These are valid:

 - ch:amd64/focal/mongodb-64: {}
 - ch:amd64/focal/ubuntu-19: {}

After running they become:

 - ch:amd64/focal/mongodb-19: {}
 - ch:amd64/focal/mongodb-64: {}
 - ch:amd64/focal/ubuntu-19: {}
 - ch:amd64/focal/ubuntu-64: {}