Prevent charm revision calling on startup
On very large deployments we do not want the charm revision worker
starting up and instantly calling the charmhub/charmstore APIs. In a HA
environment this is compounded by the fact that it does it for every
controller node. This can happen if you bounce the agents, so jujus
under load or in trouble are not helping themselves out here.

The solution here is half baked, I've left a TODO to state what the
correct solution is, for now we will just jitter the worker for the
delay so that not every node is smashing the API at the same time.

The real solution is to have a lease for the worker so that it is only
responsible for calling the API. Releasing that lease when it's
terminated. We can not add ifResponsible flag in the manifold, because
if the agent is being (hard) bounced then there is a chance that this
never runs.

I think for now, this is a happy medium.