Remove non-passing acceptance tests
In order to get the signal back from our CI tests, we have some tests
that just never passed. As the people whom wrote them are long gone and
the new preferred way is to write using the shell tests we should nuke
from orbit.

These tests will be added to a document to state that they do need
re-writing in the future, but these tests have always been flaky and we
REALLY want to improve our signal to noise ratio.