[JUJU-438] State manager worker (#13618)
* Refactor dbaccessor worker

The worker is now refactored to allow for testing and building with cgo
and no-cgo implementations. This should drastically improve workflow. It
also ensures that the cgo changes do not leak outside of this folder for
now.

More work is require for the repl, which will following in other
commits.

* Return REPL from a DBApp

Depending on the DBApp it will return a REPL based on that. For noop
DBApp returns not supported, it probably should just return nil for the
error and just ignore accessing the REPL.

* Rename noop app to unsupported app

As it's no longer just performing non-operations, it's actually erroring
out when calling a method, we should name the struct accordingly.

* StateManager worker

The following starts the implementation of passing back a state manager
to prevent direct access to the db. The db is now hidden behind a
statemanager and all access to the underlying state is through the
different managers.

* Start making the tests work again

The tests are failing because the context is leaked out and we need to
find a better way to create unit tests for them.

* Implementation of the overlord for logging

As we partition the logs for the controller into a different database we
need to manage that with a different logging overlord state.

Each model will have it's own overlord.

* Rename state manager to state manager provider

As this provides state managers to becomes confusing when it is also
called state manager.

* Scaffold the txn runner.

* Move schema applier to overlord

Move the schema applier to the overlord so that all sql statements are
in one location.

* Perform all operations in a transaction

It is expected that all queries must be running through a transaction,
that way we enable batching via explicit transaction rather than
implicit transactions.

See information about the following write up[1] and replicating the
expectations in dqlite[2].

 1. https://medium.com/@JasonWyatt/squeezing-performance-from-sqlite-insertions-971aff98eef2
 2. https://gist.github.com/SimonRichardson/232ac91c5cb49c5a0c8048a93a929bb6

* Re-write transaction builder

The transaction builder now runs all transaction runners on commit,
which then allows retries much more simpler as we can get transient
network issues.

-----------------------------------------------------------------------

Export CGO_ENABLED=0

The following exports CGO_ENABLED=0 in the make file so we don't have to
change every build target. At a later date we can update this to be
specific, but for now brute force it.

* Namespace as constants

Make the namespaced overlords constants to prevent getting an invalid
overlord.

* Allow the returning of a TxnBuilder for a more fluent API