Re-implement GetEssentialMetadata for charmhub repo using ch refresh API
To obtain the essential metadata (metadata.yaml, config.yaml and
manifest.yaml) we invoke the refresh API in batch mode. The API provides
the contents of metadata and config YAML files when instructed to do so
at request time. However, we need to map the "bases" field from
the response into a manifest instance manually.

The new implementation is not as efficient as it should be as we need to
resolve the origin of every charm via a separate call
(ResolveWithPreferredChannel repository method).

We should provide a batch-aware implementation of the above method so we
can resolve all origins via a single call.

Alternatively, we could opt to reconstruct the resolved origin from the
refresh API response (the one containing the metadata we need). However,
the logic for doing so is *very* convoluted so the decision was made to
make an extra call as opposed to duplicating the logic.