Derive the status from the application units
The following derives the status from the application units if the
application status is set to unset. This is because the application
status is no longer inherited from the units, so when we require to know
the application status, we have to inspect all the units.

This was changed to the model cache to make it more performant, except
we need to now remove the model cache to fit the new data model and to
get a new write through cache.