Require only a ping for the state pool
The state model worker would force a mongo session ping per model. This
causes a lot of noise for controllers with lots of models. In reality
all sessions for a given controller are copies of the root session. So
if the root session isn't pingable then it doesn't matter if the model
sessions are reachable, we'll fail at the system state level. We use the
catacomb dying as a way to indicate that all modelStateWorkers should
clean up and release the state.

Originally the premise was to remove the modelStateWorker, but we do
still require the modelStateWorker as that correctly removes the model
from the state pool. I don't believe there is another way around
that, without using finalizers.

As a driveby I also pushed the clock in through the manifold.