Refactor CheckProviderAPI.
To use either the new CloudEndpointChecker or the CheckProvider
interfaces when validating connectivity to the environ provided.
This is a half way step for moving away from calling the heavy
weight AllInstances simply to verify juju can still talk to the model's
cloud before a model upgrade.

The idea was to use Ping, however it's not implemented by any of the
public cloud providers.  ValidateCloudEndpoint is a wrapper around Ping
to also get the cloud endpoint which Ping requires.  It should be
implemented for all providers moving forward.  At which point the
CheckProvider interface can be removed.