Charm revision updater prevent terminal failure
If either charmhub or charmstore goes away or starts serving bad
content, like in the charmstore case, we shouldn't die on the hill.
Instead we should do our best effort to at least return some content.
This sort of async job can be thought of as a tier 2 part of juju, it's
not massively critical, but very useful UX feedback. This may change in
the future, but as it stands this shouldn't just flatout fail.

The following just changes how the errors are handled, whilst trying the
best to return the correct values and errors accordingly.

The integration test still fails for charmstore as they're serving bad
content, but it works nicely for charmhub, as the new test identifies.