Allow the ability to specify branch when deploying charms
The following is the initial work for deploying charms with channel
branches. As channel branches are non-discoverable it means that both
arch and series (platform) are required when using a branch in a
channel.

The code was pretty much wired up for this already, it was just about
exposing this via the api parameters.