Attempt to fix race condition in worker/caasunitprovisioner
I think what was happening was occasionally the parent worker was
getting another event that would cause it to remove the
applicationWorker before the test expected it, and the
c.Assert(ok, jc.IsTrue) would fail (or previously panic when the assert
wasn't there).

Now we fetch the applicationWorker earlier in the test when we know it
should be active, so the CheckKilled argument will never be nil.