Clean up error message for raft lease request
If we don't have a client yet and is expected to process a raft lease
command, we want the error (now warning) message to be less cryptic.

Dropped to a warning, as operators should be aware of the message, but
don't need to jump onto the error message itself. This should resolve
itself once a client has been connected.