Changes to image-check build subprocess.
In the previous versions of our make file the operator image building
would determine if it was needed to build the Juju bin's and call a Make
subprocess to do the building.

This has the negative effect where Make options like -j 4 would not get
passed along slowing down the Make build. This change keeps the build in
the same process to maintain the env.