Create a new raft notifier worker
The raft notifier worker isolates state away from the raft
implementation. The basic premise is that if state is bouncing, we don't
want raft bouncing. Keeping raft up prevents a lot of churn in
application leases.

The new notify proxy consumes notifications from raft and writes them to
a buffer. If a consumer to the proxy isn't listening, we just populate
the buffer. Then when the consumer is ready, it can consume the buffer.