Always set the application arch
The application arch should always be set so when in a non default
architecture environment we can always add a new unit without setting
the model constraints.

The hierarchy of needs for adding an application and therefore a unit
will first check the user passed constraints arch arguments, before
then checking the model constraints arch value. Reality we would want to
move constraints in to config, but that's an exercise for later.

The code is simple, when we create a new application, just ensure we
always have a default arch.