Add unit state data into the introspection report.
When the unit state was moved from a file in the juju directory to the controller db,
we lost the ability to easily view from the unit's machine, what the unit believes is
its current state.  Correct that now.

Potentially in the future, additional hook data would be useful.  For now, report
the hook kind.