Implement ExpireLeases.
This still needs a lot more testing. Especially around the potential for
concurrency and having multiple processing trying to extend leases and
having something decide that it also needs to expire them. However, the
basics feel like they will be 'consistent'. With the only real
discrepancy being something like a TXN that wants to remove a Holder but
the Expiry entry has already moved on, etc.

The basics are there, the data is validated, need to get some more tests
and figure out the exact interactions.