Supply a notify proxy for lease changes
The following updates raft worker to use a notify proxy so itself
doesn't have to depend on state. The idea behind this, is to keep raft
and other workers up if state is bouncing. Although this is a tough
challenge, it's something that will be worth while.

The concept is simple, provide a worker that just depends on both state
and raft, if state goes down then only bounce the new worker and state,
not raft.

The conduit between these two workers is a notifier proxy. It consumes
messages from the raft work and places them on a channel. If the
notifications aren't consumed then the notifications are left in a
poormans LRU buffer, until the consumer starts to consume.

That way we have a better uptime strategy without the downsides.