Pass macaroons from cookiejar for HTTPS requests
The following passes the macaroons for all https requests if you
logout and then login. As we rightly do not store the new password once
you've changed it, we then expect to use the macaroons from the cookie
jar. We do that correctly when passing to login, but we fail to do it
when calling the HTTPS requests.

The following fix just creates a new openFunc middleware that just adds
the macaroons if the password and the length of macaroons is empty. For
most if not all cases this should not be called, yet for uploading
charms in production environments where the password for the user has
been changed it clearly a benefit.