Allow blocking and non-blocking modes for the proxy
The proxy can work in either blocking or non-blocking modes. The
blocking mode is ideal to ensure the queries (claimed, expiries) are
correctly sent through. In production we don't actually care atm. We can
then switch to blocking mode when the underlying raft worker cares.

The design of the notify proxy is for us to allow dropping of queries if
nobody is consuming the other side and when they do reconnect they can
pick off where they left off. The caveat being as long as it's with in
the buffer range.

Work on the buffer size will come later, but at 512, that should be
enough for some environments and we can always use the trapdoor as the
fallback.