Fixes a bug for zero credentials in finalise.
When finalising a k8s clouds we try and find credentials for clouds that
we have auto filled so that we may check the storage of the cluster. For
clouds added by the user we can't do these checks nor do we have to.
This adds a check to see if we can find credentials and if not we don't
perform any finalise actions on k8s clouds.

Side note this code is become horrible to test and needs to be broken
out away from the main provider so that it can be sanity tested.