Ensure the actual charm origin is saved.
Usually the initial and actual charm origin will be the same.  However in the case of refresh with --switch, from a local or charmstore charm to a charmhub charm, the ID and hash will be missing.  The ID is essential for when a charm might be renamed in the store.  It allows for juju to have a method to identify a charm independent of the name and refresh to the correct thing when a name changes.