Simplify facade construction
As a forward step towards a more simplified facade setup, which is
required for the data modelling changes. The following cleans up how a
facade can be constructed. A facade now only has one type signature. It
must take a facade.Context and return an interface{} or error.

In following commits/PRs we should be able to remove a LOT of the
reflect setup for registering a facade. The general purpose is to allow
us to inject a state when required using the correct context.

There will be lots of failed tests in this first commit, so we'll get CI
to run them.