Compute base channel for centos
The following allows centos charm authors to use just the series number
for the series instead of centos7. The code is backwards compatible with
older style series naming. Because we pushed this to the edges for
normalisation, the data stored in the database is correct. It
reports it as centos7 and we don't leak the implementation information
outside of the charmhub repository.