Hide debug.Stack calls behind a feature flag
We really don't want to be calling debug.Stack for every call to
everything. It might be negligible from a small model point of view, but
for HUGE models we REALLY don't want to be doing this busy work.

For every call, it makes a minimum 1024 byte slice, if it overflows it
will grow larger and larger. As juju is doing this all the time, this
will add up.

I've not removed this, just hidden it behind either a feature flag or a
trace enable call.

I suspect we should get a speed up here.