Remove facade reflection for construction
The following removes all the facade reflection when building a new
facade. As we no longer support multiple ways of creating a facade, we
can force everything to be typed for construction.

Additionally we're inverting the facade registration. Facades should
not expose their internal details outside of the package. Instead that's
an internal implementation detail. We use the visitor pattern to invert
the dependency. This means we DON'T have to be generic about how we
construct each facade. We can prevent public construction methods from
being public and provide a much tighter API.