Replace charm Source in listOneCharmResource getCharmRespository call.
While technically correct to call getCharmRespository with a CharmHub
source, use the source in the normalized origin to avoid a potential
future mistakes. ListResources technically can only be run on CharmHub
charms, it errors if not. Charm store charms a very different code path
is used starting in the charm-resources command file.