cloudconfig: Skip IPv6 link-local in Netplan
Netplan doesn't allow IPv6 link-local addresses in its configuration,
specifically, it doesn't seem to support the '%interface' or '%id'
syntax which is required on such addresses.

As far as I can tell, there is no such limitation to ifupdown which will
then pass that value directly to /etc/resolv.conf, so this fix only
applies to the Netplan generator.

Bug: https://bugs.launchpad.net/juju/+bug/1883701

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>