Update the version of juju/errors and errors usage
With this commit we are updating our version of juju/errors and the
associated usage of how we go about doing errors. This error now allows
us to use proper support for the new constructs of Is, As and Unwrap.

A lot of the changes in this commit are to address functional uses of
our errors library and specifically allowing us to deal with chains of
errors and know longer rely on type checking.

Notable changes:
- Turned var errors into const errors where it could be done.
- Updated the juju/errors library.
- Fixed a re-entrant bug in the aws provider for invalidating credentials.
- Fixed a bug in the gui apiserver code that was testing for the wrong
  filename value in errors.
- Removed testing checks that assert the functionality of juju/errors
  tracing output. This is well tested from within the package so it
  makes no sense to double up on testing in Juju.
- Change a lot of the IsErrType function check to use standard errors.Is

In future updates of errors usage in Juju we should actively avoid
trying to errors.Wrap() 2 or more errors together and opt for a stable
chain of linked errors.