Update the openstack base bundle used in the deploy test.
The current bundle use the lxd charm. With the change from charmstore to
charmhub, compatibility was broken for the lxd charm. It changed from a
subordinate to a principle charm. Adding a new version of the bundle
and update the test to wait for idle appropriately.

This bundle should be replaced for good moving forward as it uses
Xenial. There is a core kubernetes bundle, however kubernetes's models
do not cleaning tear down today.