Ensure centos/7 is normalised at the edge
The following is a bit convoluted as we need to ensure that we normalise
the platform series at the edge. By edge we mean we try and normalise
at the call to charmhub and keep the series "pure" in juju.

The modelling of series in juju is wrong it should be a tuple base of
(os/version) and not just bionic, focal etc.