Discard syslog for k8s models
The following discards the syslog logger for k8s models. The idea being
that syslog isn't always available in every container, so we should
prevent it for now. In the future, it would be better for k8s models to
not speak to a local syslog, but a remote one, like the log forwarder.

Until that's setup just discard any logs for now. As this is a
optimisation for production issues we can ignore this for normal use
cases.