Use dry-run for repository charms only, not local.
Dry run with a local charm would require uploading the charm to
determine a series etc. As well, the revision of the local charm
would increase each time dry run was used. Neither is desirable
for dry run at this time.

Move the dry run check back into repository charm's PrepareAndDeploy.
CharmInfo in deploy requires an uploaded charm which we do not want.

Not uploading the charm on dry run limits the full set of checks to be
done such as validation of the config used. This can be addressed if and
when the same bug for bundle dry run is resolved.