Make no assumptions about a docs existence when used as a txn assert.
When updating a pending resource to deploy or refresh an application, do not assume that because a resource doc does not exist, that neither does the store version of the doc. This assumption causes a state changing too quickly error and prevents application refresh.

Found as part of LP 1968931, #charmhub docs were inserted by the charmrevisionupdater. The deployed revision of the charm had no resources, but the newer revision did.

Update log messages which did not indicate the true doc queried for to improve debugging.